fix(precios): CLP sin decimales y separador de miles en el front
- BD: workshops.price -> INTEGER (migracion 000021 + schema base init/qa)
- Backend PHP: cast ::int en queries de lectura y round() en escrituras; Workshop model cast a integer
- Front: formatPrice unificado con separador de miles (punto) en todos los displays; inputs de precio sin decimales

+++ apps/api-php/database/migrations/2024_01_01_000004_create_workshops.php
@@ -16,7 +16,7 @@ return new class extends Migration
             description        TEXT,
             type               workshop_type NOT NULL DEFAULT 'workshop',
             modality           workshop_modality NOT NULL DEFAULT 'in-person',
-            price              NUMERIC(10,2) NOT NULL DEFAULT 0,
+            price              INTEGER NOT NULL DEFAULT 0,
             currency           VARCHAR(3) NOT NULL DEFAULT 'CLP',
             capacity           INTEGER,
             location           VARCHAR(255),
